requirements.txt
Flask==3.0.3
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
 